requirements.txt
flask
waitress
python-dateutil
PasteDeploy
keystonemiddleware
oslo.policy
